+++ phillydb/search_queries.py
@@ -4,9 +4,20 @@ from phillydb.exceptions import (
     SearchTypeNotImplementedError,
 )
 
+def search_method_sql(search_to_match, search_method):
+    if search_method == "contains":
+        return f"%{search_to_match}%"
+    elif search_method == "starts with":
+        return f"{search_to_match}%"
+    elif search_method == "ends with":
+        return f"%{search_to_match}"
+    elif search_method == "equals":
+        return search_to_match
+    else:
+        raise SearchMethodNotImplementedError(search_method)
 
 def construct_search_query(
-    search_query, search_type, search_method="contains",
+    search_to_match, search_type, search_method="contains",
 ):
     """
     Takes in a search query and method of searching, and constructs a SQL subquery
@@ -15,7 +26,7 @@ def construct_search_query(
 
     Parameters
     ----------
-    search_query: str
+    search_to_match: str
     search_method: str
         One of: ['contains', 'starts with', 'ends with', 'equals']
     search_type: str
@@ -24,44 +35,34 @@ def construct_search_query(
             ]
     """
     print("Loading data...")
-    search_query = search_query.upper()
-    if search_method == "contains":
-        search_query = f"%{search_query}%"
-    elif search_method == "starts with":
-        search_query = f"{search_query}%"
-    elif search_method == "ends with":
-        search_query = f"%{search_query}"
-    elif search_method == "equals":
-        search_query = search_query
-    else:
-        raise SearchMethodNotImplementedError(search_method)
+    search_to_match = search_method_sql(search_to_match.upper(), search_method)
 
     if search_type == "owner":
         opa_account_numbers_sql = f"""
             SELECT opp.parcel_number FROM opa_properties_public opp 
                 WHERE (
-                    owner_1 LIKE '{search_query}' OR owner_2 LIKE '{search_query}'
+                    owner_1 LIKE '{search_to_match}' OR owner_2 LIKE '{search_to_match}'
                 )
             UNION ALL
             SELECT opa_account_num FROM business_licenses bl, opa_properties_public opp
                 WHERE bl.opa_account_num = opp.parcel_number
-                AND licensetype = 'Rental' and licensestatus = 'Active'
+                AND licensetype = 'Rental'
                 AND (
-                    legalname LIKE '{search_query}' OR business_name LIKE '{search_query}'
+                    legalname LIKE '{search_to_match}' OR business_name LIKE '{search_to_match}'
                 )
         """
     elif search_type == "mailing_address":
         opa_account_numbers_sql = f"""
         SELECT opp.parcel_number FROM opa_properties_public opp
             WHERE (
-                mailing_address_1 LIKE '{search_query}' 
-                OR mailing_street LIKE '{search_query}'
+                mailing_address_1 LIKE '{search_to_match}' 
+                OR mailing_street LIKE '{search_to_match}'
             )
         """
     elif search_type == "location_by_owner":
         opa_account_numbers_sql = f"""
          SELECT opp2.parcel_number FROM opa_properties_public opp1, opa_properties_public opp2 
-            WHERE opp1.location LIKE '{search_query}' AND (
+            WHERE opp1.location LIKE '{search_to_match}' AND (
                 opp1.owner_1 = opp2.owner_1 OR opp1.owner_2 = opp2.owner_2
                 OR opp1.owner_1 = opp2.owner_2 OR opp1.owner_2 = opp2.owner_1
             )
@@ -69,7 +70,7 @@ def construct_search_query(
     elif search_type == "location_by_mailing_address":
         opa_account_numbers_sql = f"""
          SELECT opp2.parcel_number FROM opa_properties_public opp1, opa_properties_public opp2 
-            WHERE opp1.location LIKE '{search_query}' AND ( 
+            WHERE opp1.location LIKE '{search_to_match}' AND ( 
                 opp1.mailing_street = opp2.mailing_street OR (
                     opp1.mailing_address_1 = opp2.mailing_address_1 
                     AND opp1.mailing_address_2 = opp2.mailing_address_2