+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 Development
 -----------
 
-* Placeholder
+* Add another query that can be used for autocomplete `query_by_single_str_column`.
+* Remove constraint for owner search on rental licenses that are active.
+* Rename PhiladelphiaDataTable to PhiladelphiaCartoDataTable
 
 0.0.2
 -----